tests finalization for yg and gkg
